+++ src/cameras/2d/effects/index.js
@@ -1,0 +1,17 @@
+/**
+ * @author       Richard Davey <rich@photonstorm.com>
+ * @copyright    2018 Photon Storm Ltd.
+ * @license      {@link https://github.com/photonstorm/phaser/blob/master/license.txt|MIT License}
+ */
+
+/**
+ * @namespace Phaser.Cameras.Scene2D.Effects
+ */
+
+module.exports = {
+
+    Fade: require('./Fade'),
+    Flash: require('./Flash'),
+    Shake: require('./Shake')
+
+};